gitserver: Move authenticated clone URL into gitserver package (#57105)
This function generates a clone URL with the credentials from the external service config. It is currently only used in gitserver and for sensitivity reasons should not be used anywhere else.
So to make that clearer, and also to organize the code a bit better, we move it into a subpackage of the server package.
