+++ cmd/gitserver/server/cloneurl/BUILD.bazel
@@ -1,0 +1,54 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("//dev:go_defs.bzl", "go_test")
+
+go_library(
+    name = "cloneurl",
+    srcs = ["clone_url.go"],
+    importpath = "github.com/sourcegraph/sourcegraph/cmd/gitserver/server/cloneurl",
+    visibility = ["//visibility:public"],
+    deps = [
+        "//internal/conf/reposource",
+        "//internal/database",
+        "//internal/encryption/keyring",
+        "//internal/extsvc",
+        "//internal/extsvc/awscodecommit",
+        "//internal/extsvc/azuredevops",
+        "//internal/extsvc/bitbucketcloud",
+        "//internal/extsvc/bitbucketserver",
+        "//internal/extsvc/gerrit",
+        "//internal/extsvc/github",
+        "//internal/extsvc/github/auth",
+        "//internal/extsvc/gitlab",
+        "//internal/extsvc/gitolite",
+        "//internal/extsvc/pagure",
+        "//internal/extsvc/perforce",
+        "//internal/extsvc/phabricator",
+        "//internal/httpcli",
+        "//internal/types",
+        "//lib/errors",
+        "//schema",
+        "@com_github_sourcegraph_log//:log",
+    ],
+)
+
+go_test(
+    name = "cloneurl_test",
+    srcs = ["clone_url_test.go"],
+    embed = [":cloneurl"],
+    deps = [
+        "//internal/database/dbmocks",
+        "//internal/extsvc/awscodecommit",
+        "//internal/extsvc/azuredevops",
+        "//internal/extsvc/bitbucketcloud",
+        "//internal/extsvc/bitbucketserver",
+        "//internal/extsvc/gerrit",
+        "//internal/extsvc/github",
+        "//internal/extsvc/gitlab",
+        "//internal/extsvc/perforce",
+        "//internal/extsvc/phabricator",
+        "//internal/timeutil",
+        "//schema",
+        "@com_github_google_go_cmp//cmp",
+        "@com_github_sourcegraph_log//logtest",
+    ],
+)
