+++ cmd/gitserver/server/cloneurl/clone_url_test.go
@@ -1,4 +1,4 @@
-package repos
+package cloneurl
 
 import (
 	"context"
