+++ cmd/gitserver/shared/BUILD.bazel
@@ -14,6 +14,7 @@ go_library(
     deps = [
         "//cmd/gitserver/server",
         "//cmd/gitserver/server/accesslog",
+        "//cmd/gitserver/server/cloneurl",
         "//cmd/gitserver/server/perforce",
         "//internal/actor",
         "//internal/api",
@@ -46,7 +47,6 @@ go_library(
         "//internal/jsonc",
         "//internal/observation",
         "//internal/ratelimit",
-        "//internal/repos",
         "//internal/requestclient",
         "//internal/service",
         "//internal/trace",
