+++ internal/repos/BUILD.bazel
@@ -8,7 +8,6 @@ go_library(
         "azuredevops.go",
         "bitbucketcloud.go",
         "bitbucketserver.go",
-        "clone_url.go",
         "conf.go",
         "discoverable_sources.go",
         "doc.go",
@@ -41,7 +40,6 @@ go_library(
         "sync_worker.go",
         "syncer.go",
         "testing.go",
-        "util.go",
     ],
     importpath = "github.com/sourcegraph/sourcegraph/internal/repos",
     visibility = ["//:__subpackages__"],
@@ -131,7 +129,6 @@ go_test(
         "azuredevops_test.go",
         "bitbucketcloud_test.go",
         "bitbucketserver_test.go",
-        "clone_url_test.go",
         "gerrit_test.go",
         "github_test.go",
         "gitlab_test.go",
@@ -150,7 +147,6 @@ go_test(
         "store_test.go",
         "sync_worker_test.go",
         "syncer_test.go",
-        "util_test.go",
     ],
     data = glob(["testdata/**"]),
     embed = [":repos"],
@@ -172,15 +168,12 @@ go_test(
         "//internal/extsvc",
         "//internal/extsvc/auth",
         "//internal/extsvc/awscodecommit",
-        "//internal/extsvc/azuredevops",
         "//internal/extsvc/bitbucketcloud",
         "//internal/extsvc/bitbucketcloud/testing",
         "//internal/extsvc/bitbucketserver",
-        "//internal/extsvc/gerrit",
         "//internal/extsvc/github",
         "//internal/extsvc/gitlab",
         "//internal/extsvc/gitolite",
-        "//internal/extsvc/perforce",
         "//internal/extsvc/phabricator",
         "//internal/github_apps/types",
         "//internal/gitserver",
